+++ src/bindings/scripts/scripts/zone/zulgurub/boss_hazzarah.cpp
@@ -72,7 +72,7 @@ struct TRINITY_DLL_DECL boss_hazzarahAI : public ScriptedAI
             //We will summon 3 illusions that will spawn on a random gamer and attack this gamer
             //We will just use one model for the beginning
             Unit* target = NULL;
-            for(int i = 0; i < 3;i++)
+            for(uint8 i = 0; i < 3;i++)
             {
                 target = SelectUnit(SELECT_TARGET_RANDOM,0);
                 if(!target)
